Move all classes into the classes folder
