+++ node/events.js
@@ -11,11 +11,11 @@ const fs = require('fs'),
   rp = require('request-promise'),
   mysql = require('mysql'),
   sqlite3 = require('sqlite3').verbose(),
-  Hosts = require('./Hosts'),
-  WsTokens = require('./WsTokens'),
+  Hosts = require('./classes/Hosts'),
+  WsTokens = require('./classes/WsTokens'),
   HostEvents = require('./classes/HostEvents'),
-  Terminals = require('./Terminals'),
-  VgaTerminals = require("./VgaTerminals"),
+  Terminals = require('./classes/Terminals'),
+  VgaTerminals = require("./classes/VgaTerminals"),
   AllowedProjects = require("./classes/AllowedProjects"),
   hosts = null,
   hostEvents = null,
